requirements.txt
chainlit
openai
python-dotenv